requirements.txt
blinker==1.9.0
click==8.3.0
Faker==38.0.0
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
lorem-text==3.0
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pillow==12.0.0
Pygments==2.19.2
rich==14.2.0
shellingham==1.5.4
SQLAlchemy==2.0.44
typer==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
